# backend/src/main/java/dev/salman/backend/features/auth/dto/LoginRequest.java
package dev.salman.backend.features.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class LoginRequest {
    private String username;
    private String password;
}