# backend/src/main/java/dev/salman/backend/features/auth/dto/RegisterRequest.java
package dev.salman.backend.features.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RegisterRequest {
    private String username;
    private String password;
    private String role; // ROLE_ADMIN, ROLE_DOCTOR, ROLE_RECEPTIONIST
    private String fullName;
}