# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TokyoClinic OS API",
    "description": "REST API contract specification for patient intake, scheduling, consultation, and live queue flows.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate Operator Node",
        "description": "Verifies staff credentials and returns a secure HttpOnly session cookie containing the JWT.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session established successfully.",
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-Only secure cookie containing the authentication token.",
                "schema": {
                  "type": "string",
                  "example": "token=eyJhbGc...; Path=/api; Max-Age=86400; HttpOnly; Secure; SameSite=Strict"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials provided."
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "summary": "Register Staff Node",
        "description": "Creates a new clinic staff account. Restricted to ROLE_ADMIN.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Staff registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Username already taken or invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions (requires Admin role)."
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get Current Profile",
        "description": "Returns details of the currently authenticated operator.",
        "responses": {
          "200": {
            "description": "Operator profile details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or missing JWT cookie."
          }
        }
      }
    },
...
    "/appointments": {
      "post": {
        "summary": "Book Appointment",
        "description": "Schedules a doctor consultation. Rejects booking if there is a scheduling conflict within a 30-minute window. Uses optimistic locking.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment booked successfully. Async SMS alert dispatched.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400": {
            "description": "Validation errors or scheduling conflict."
          }
        }
      }
    },
    "/appointments/doctor/{doctorId}": {
      "get": {
        "summary": "Get Doctor Schedule",
        "description": "Lists appointments assigned to the doctor, ordered by time ascending.",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assigned appointments array.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{id}/status": {
      "put": {
        "summary": "Update Appointment Status",
        "description": "Transitions slot state (e.g. to CANCELLED). Rejects transition if slot is already cancelled or completed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["SCHEDULED", "CANCELLED", "COMPLETED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment slot updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "receptionist"
          },
          "password": {
            "type": "string",
            "example": "recept123"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "role", "fullName"],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["ROLE_ADMIN", "ROLE_DOCTOR", "ROLE_RECEPTIONIST"]
          },
          "fullName": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          }
        }
      },
...
      "AppointmentRequest": {
        "type": "object",
        "required": ["patientId", "doctorId", "appointmentTime"],
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}